fix PR comments 3
