+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeList/hooks/useInputCodeListNameErrorMessage.ts
@@ -1,7 +1,7 @@
 import { FileNameValidationResult } from '@studio/pure-functions';
 import { useTranslation } from 'react-i18next';
 
-export function useValidateInputCodeListName() {
+export function useInputCodeListNameErrorMessage() {
   const { t } = useTranslation();
 
   type FileNameValidationErrorResult = Exclude<
